refactor: move demo to its own folder
