+++ package.json
@@ -24,7 +24,7 @@
     "test:ui": "SLOW_MO=800 playwright test --ui --headed --workers=1",
     "test:dev:ui": "E2E_DEV=1 SLOW_MO=800 playwright test --ui --headed --workers=1",
     "cron:backup": "tsx scripts/cron-backup.ts",
-    "demo": "tsx scripts/demo.ts"
+    "demo": "tsx demo/index.ts"
   },
   "dependencies": {
     "@dnd-kit/core": "^6.3.1",
