+++ tsconfig.json
@@ -30,5 +30,5 @@
     ".next/dev/types/**/*.ts",
     "**/*.mts"
   ],
-  "exclude": ["node_modules", "e2e"]
+  "exclude": ["node_modules", "e2e", "demo"]
 }
